Add end session

+++ app.py
@@ -70,12 +70,12 @@ class StatusBarApp(rumps.App):
         sessions[self.current_session] = url_list
         self.write_sessions(sessions)
         print("----Tabs recorded----")
+        self.update_all_sessions()
 
     @rumps.clicked("Start session")
     def record_tabs(self, _):
         """Starts a new session, records tabs, and calls update_tabs to refresh the menu to reflect this"""
         response = rumps.Window(
-            default_text="my cool session",
             cancel="Cancel",
             title="Enter a session name",
             dimensions=(300,20)
@@ -94,6 +94,10 @@ class StatusBarApp(rumps.App):
         self.current_session = session_name
         self.update_tabs()
 
+    def end_session(self, _):
+        self.current_session = ""
+        self.update_all_sessions()
+
     def get_session_names(self, tabs_file):
         session_names = []
         with open(tabs_file, "r") as f:
@@ -111,23 +115,25 @@ class StatusBarApp(rumps.App):
             self.delete_menu.append(delete_item)
 
     def update_all_sessions(self):
-        print("update all sessions was called")
         session_names = self.get_session_names(self.tabs_file)
         load_menu = rumps.MenuItem("Load session")
         delete_menu = rumps.MenuItem("Delete session")
         for name in session_names:
+            if name == self.current_session: continue
             load_item = rumps.MenuItem(name, callback=self.load_session)
             delete_item = rumps.MenuItem(name, callback=self.delete_session)
             load_menu.add(load_item)
             delete_menu.add(delete_item)
         self.menu.clear()
-        self.menu.add(rumps.MenuItem("Start session", callback=self.record_tabs))
+        if self.current_session:
+            self.menu.add(rumps.MenuItem(self.current_session))
+        if self.current_session:
+            self.menu.add(rumps.MenuItem("End session", callback=self.end_session))
+        else:
+            self.menu.add(rumps.MenuItem("Start session", callback=self.record_tabs))
         self.menu.add(load_menu)
         self.menu.add(delete_menu)
         self.menu.add(rumps.MenuItem("Quit", callback=rumps.quit_application))
-        if self.current_session:
-                self.menu.add(rumps.MenuItem(self.current_session))
-                print("added current_session to menu")
 
     def load_session(self, var):
         """loads a session from the text file, and updates menu to reflect this"""
@@ -135,7 +141,6 @@ class StatusBarApp(rumps.App):
         websites = session_data[var.title]
         subprocess.check_output(["open"] + websites)
         self.current_session = var.title
-        print
         self.update_all_sessions()
 
     def delete_session(self, var):
